Fix many many warnings and other issues thrown up by DAI_GNUC_PRINTF().
This also fixes the bug which allowed players to type printf conversion
specifications and have them converted by the server (this was also the case
for scripts and object names, etc). Typing %n would crash the server! Now
typing %n prints %n.

Functions which take a format argument *must* have a string literal format
argument!

+++ server/src/common/shstr_ng.c
@@ -430,7 +430,7 @@ static void ss_find_totals(int *entries, int *refs, int *links, int what)
         *refs += ss->refcount;
         *links += probes;
         if(what & SS_DUMP_TOTALS)
-            LOG(llevSystem, "%4d -- %4d refs '%s' (%d probes)\n", i, (ss->refcount), ss->string, probes);
+            LOG(llevSystem, "%4d -- %4d refs '%s' (%d probes)\n", (int)i, (int)(ss->refcount), ss->string, probes);
     }
 }
